[HUMAN App] refactor: service sign up (#3184)
