+++ packages/apps/human-app/frontend/src/modules/signin/worker/use-sign-in.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
-import { useSignInMutation } from '@/modules/worker/services/sign-in/sign-in';
 import { browserAuthProvider } from '@/shared/contexts/browser-auth-provider';
-import { type SignInDto } from '@/modules/worker/services/sign-in/types';
+import { useSignInMutation } from './use-sign-in-mutation';
+import { type SignInDto } from './schemas';
 
 export function useSignIn() {
   const {
