--- packages/apps/human-app/frontend/src/modules/worker/services/sign-in/types.ts
@@ -1,9 +1,0 @@
-import { type z } from 'zod';
-import {
-  type signInDtoSchema,
-  type signInSuccessResponseSchema,
-} from './schema';
-
-export type SignInDto = z.infer<typeof signInDtoSchema>;
-
-export type SignInSuccessResponse = z.infer<typeof signInSuccessResponseSchema>;
