+++ packages/apps/human-app/frontend/src/shared/schemas/index.ts
@@ -1,2 +1,3 @@
 export * from './validate-address-schema';
 export * from './url-domain-schema';
+export * from './auth-tokens-schema';
